+++ Editor/Unity.Build.Classic.Private/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 ﻿using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleTo("Unity.Platforms.Android.Build")]
-[assembly: InternalsVisibleTo("Unity.Platforms.Web.Build")]
-[assembly: InternalsVisibleTo("Unity.Platforms.Windows.Build")]
-[assembly: InternalsVisibleTo("Unity.Platforms.iOS.Build")]
-[assembly: InternalsVisibleTo("Unity.Platforms.macOS.Build")]
+
+// Per platform access
+[assembly: InternalsVisibleTo("Unity.Build.Android.Classic")]
+[assembly: InternalsVisibleTo("Unity.Build.iOS.Classic")]
+[assembly: InternalsVisibleTo("Unity.Build.macOS.Classic")]
+[assembly: InternalsVisibleTo("Unity.Build.Windows.Classic")]