+++ Editor/Unity.Build/AssemblyInfo.cs
@@ -5,3 +5,6 @@
 [assembly: InternalsVisibleTo("Unity.Build.Editor")]
 [assembly: InternalsVisibleTo("Unity.Build.Internals")]
 [assembly: InternalsVisibleTo("Unity.Build.Tests")]
+
+// For ICustomBuildComponentConstructor
+[assembly: InternalsVisibleTo("Unity.Build.Android")]